feat: add the SHAKE extendable-output functions
Instead of creating a separate category for hash-based functions such as
Message Authentication Codes (MACs), these primitives will be solely
assessed on the hash functions they take as inputs through a separate
function in the standards module.

For example, the security of a HMAC primitive that is used with the
SHA256 hash function will be assessed based on the digest length of the
hash function which here will result in the message authentication tag.
By most standards these can be twice shorter for than for applications
requiring collision resistance.

Resolves: https://github.com/tshakalekholoane/wardstone/issues/22.