refactor: lower the security of SHA-1
Per page 8 of NIST SP 800-107, the SHA-1 is believe to provide less
security than its digest length suggest. While this is still fine for
first pre-image resistance, it can be misleading for applications
requiring collision resistance and second pre-image resistance and thus
a decision has been made here to use the conservative value of 105 that
appears in the same publication.