fix: update SHA1 security
The estimate for security provided by SHA1 proved overly conservative
even though SHA1 provides less security than its digest length would
lead. This enables the SHA1 algorithm to pass for pre-image resistance
for NIST as documented on p. 56 but still fail in standards such as the
BSI where it is forbidden even if there is no evidence that is property
is broken.