refactor: update elliptic curve display strings
The non-uniformity of elliptic curve group names makes it slightly
difficult to interpret output involving such curves.

For example, previously if the user had a certificate that used the
prime256v1 curve, they might have gotten the following report when
assessing its compliance against NIST guidance.

  ...
  signature algorithm: got prime256v1, want nistp256
  ok: ./crates/cmd/src/testing/certificates/ecc_prime256v1.pem

While this is probably fine, it can be further improved to indicate that
this refers to the same curve.

  ...
  signature algorithm: got nistp256, prime256v1, or secp256r1, want nistp256, prime256v1, or secp256r1
  ok: ./crates/cmd/src/testing/certificates/ecc_prime256v1.pem

This way they do not have to reconcile the names especially if they then
decide to assess the same curve against another standard i.e., BSI which
prefers to call it by the name secp192r1.